backend/requirements.txt
Django==3.1
django-cors-headers==3.6.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.5
gunicorn==20.0.4
Pillow
